Move check_running to misc module
